requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.0
docker==7.1.0
psycopg2-binary==2.9.10
sse-starlette==2.2.1
python-dotenv==1.1.0
